+++ packages/@leight/job/src/sdk/Schema.ts
@@ -4,7 +4,14 @@
  So, please, DO NOT modify this file as it would get re-generated and you would be f*cked up.
  */
 import {FilterSchema}       from "@leight/filter";
-import {PrismaSchema}       from "@leight/prisma";
+import {
+    JobOptionalDefaultsSchema,
+    JobOrderByWithRelationInputSchema,
+    JobPartialSchema,
+    JobSchema as $EntitySchema,
+    JobWhereInputSchema,
+    JobWhereUniqueInputSchema
+}                           from "@leight/prisma";
 import {
     ParamsSchema,
     QuerySchema
@@ -14,6 +21,12 @@ import {WithIdentitySchema} from "@leight/source";
 import {z}                  from "zod";
 import {JobSchemaOverride}  from "../schema";
 
+export type IJobWhereSchema = typeof JobWhereSchema;
+export type IJobWhere = z.infer<IJobWhereSchema>;
+export type IJobWhereUniqueSchema = typeof JobWhereUniqueSchema;
+export type IJobWhereUnique = z.infer<IJobWhereUniqueSchema>;
+export type IJobOrderBySchema = typeof JobOrderBySchema;
+export type IJobOrderBy = z.infer<IJobOrderBySchema>;
 export type IJobSchema = typeof JobSchema;
 export type IJob = z.infer<IJobSchema>;
 export type IJobCreateSchema = typeof JobCreateSchema;
@@ -29,19 +42,22 @@ export type IJobSort = z.infer<IJobSortSchema>;
 export type IJobQuerySchema = typeof JobQuerySchema;
 export type IJobQuery = z.infer<IJobQuerySchema>;
 
-export const JobSchema       = PrismaSchema.JobSchema.merge(JobSchemaOverride);
-export const JobCreateSchema = PrismaSchema.JobOptionalDefaultsSchema;
-export const JobPatchSchema  = PrismaSchema.JobPartialSchema.merge(WithIdentitySchema);
-export const JobFilterSchema = z.union([
-    PrismaSchema.JobWhereInputSchema,
-    PrismaSchema.JobWhereUniqueInputSchema,
+export const JobWhereSchema       = JobWhereInputSchema;
+export const JobWhereUniqueSchema = JobWhereUniqueInputSchema;
+export const JobOrderBySchema     = JobOrderByWithRelationInputSchema;
+export const JobSchema            = $EntitySchema.merge(JobSchemaOverride);
+export const JobCreateSchema      = JobOptionalDefaultsSchema;
+export const JobPatchSchema       = JobPartialSchema.merge(WithIdentitySchema);
+export const JobFilterSchema      = z.union([
+    JobWhereSchema,
+    JobWhereUniqueSchema,
     FilterSchema,
 ]);
-export const JobParamSchema  = ParamsSchema;
-export const JobSortSchema   = z.object({
+export const JobParamSchema       = ParamsSchema;
+export const JobSortSchema        = z.object({
     started: SortOrderSchema
 });
-export const JobQuerySchema  = QuerySchema({
+export const JobQuerySchema       = QuerySchema({
     filterSchema: JobFilterSchema,
     sortSchema:   JobSortSchema,
     paramsSchema: JobParamSchema,