+++ packages/@leight/prisma/src/index.ts
@@ -1,17 +1,14 @@
 import {Prisma} from "@prisma/client";
 
-export * from "./context";
-
 /**
  * Symbol for container usage of Prisma.
  */
 export const $PrismaClient = Symbol.for("@leight/prisma/PrismaClient");
-export {
-    PrismaClient,
-}        from "@prisma/client";
-
-export * as PrismaSchema from "./schema";
 
 export const decimalOf = (value: any): number => {
     return (new Prisma.Decimal(value)).toNumber();
 };
+
+export {PrismaClient} from "@prisma/client";
+export *              from "./context";
+export *              from "./schema";