+++ packages/@leight/file-server/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @leight/file-server
 
+## 0.5.53
+
+### Patch Changes
+
+- Improved generator
+- Updated dependencies
+    - @leight/prisma@0.5.61
+    - @leight/file@0.5.52
+
 ## 0.5.52
 
 ### Patch Changes