+++ packages/@leight/prisma/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leight/prisma
 
+## 0.5.61
+
+### Patch Changes
+
+- Improved generator
+
 ## 0.5.60
 
 ### Patch Changes